# openapi.yml
openapi: 3.0.3
info:
  title: ClubManager API
  description: >
    Specification for the ClubManager API,
    used by web and mobile clients to interact with the server.
  version: 0.1.0
...
components:
  securitySchemes:
  # I'm not sure if we are going to need BasicAuth depends how we handle Cognito
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: [] 
paths:
...
  /api/event/new:
    post:
      tags:
        - event
      summary: Create a new event
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/event/{eventID}:
    parameters:
      - name: eventID
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
        - event
      summary: Get this event
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    patch:
      tags:
        - event
      summary: Update this event
      description: >
        The request body needs only contain the fields with data to be updated.
        If data in a field should not be updated, that field can be omitted from
        the request body.
        <br><br>
        This endpoint requires an `Authorization: Bearer` header with a JWT ID token.
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    delete:
      tags:
        - event
      summary: Delete this event
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/event/list:
    get:
      tags:
        - event
      summary: Get all events
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      parameters:
        - name: Range
          in: header
          description: |
            Optional: filter events by a specific timeframe.
            Format is one of:
            - `YYYYMMDD-YYYYMMDD`
            - `YYYYMMDD-`
            - `-YYYYMMDD`
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        206:
          description: Partial Content
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/eventType/new:
    post:
      tags:
        - eventType
      summary: Create a new event type
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/eventType/{eventTypeID}:
    parameters:
      - name: eventTypeID
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
        - eventType
      summary: Get this event type
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    patch:
      tags:
        - eventType
      summary: Update this event type
      description: >
        The request body needs only contain the fields with data to be updated.
        If data in a field should not be updated, that field can be omitted from
        the request body.

        **Note:** This endpoint is used to mark an event type as inactive, rather
        than using a `DELETE` endpoint to remove it entirely.
        
        <br><br>
        This endpoint requires an `Authorization: Bearer` header with a JWT ID token.
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/eventType/list:
    get:
      tags:
        - eventType
      summary: Get all event types
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO

  # job
  /api/job/new:
    post:
      tags:
        - job
      summary: Create a new job
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/job/{jobID}:
    parameters:
      - name: jobID
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
        - job
      summary: Get this job
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    patch:
      tags:
        - job
      summary: Update this job
      description: >
        The request body needs only contain the fields with data to be updated.
        If data in a field should not be updated, that field can be omitted from
        the request body.

        **Note 1:** This endpoint is used to verify work points.
        
        **Note 2:** While most fields require admin permissions
        to update, both *1)* assigning the job to a member or paid laborer and
        *2)* checking in to the job can be done by a member as well.
        
        <br><br>
        This endpoint requires an `Authorization: Bearer` header with a JWT ID token.
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    post:
      tags:
        - job
      summary: Clone this job
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    delete:
      tags:
        - job
      summary: Delete this job
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
...
  /api/eventJob/new:
    post:
      tags:
        - eventJob
      summary: Create a new event-job instance
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
  /api/eventJob/{eventJobID}:
    parameters:
      - name: eventJobID
        in: path
        required: true
        schema:
          type: integer
    get:
      tags:
        - eventJob
      summary: Get this event-job instance
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    patch:
      tags:
        - eventJob
      summary: Update this event-job instance
      description: >
        The request body needs only contain the fields with data to be updated.
        If data in a field should not be updated, that field can be omitted from
        the request body.
        
        <br><br>
        This endpoint requires an `Authorization: Bearer` header with a JWT ID token.
      requestBody:
        content:
          application/json:
            schema:
              # TODO
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO
    delete:
      tags:
        - eventJob
      summary: Delete this event-job instance
      description: "This endpoint requires an `Authorization: Bearer` header with a JWT ID token."
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                # TODO
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                # TODO
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                # TODO
        404:
          description: Not Found
          content:
            application/json:
              schema:
                # TODO
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                # TODO

  # billing
...